# docs/openapi/openapi.yaml
basePath: /
definitions:
...
  dev.listAgentInteractionsResponse:
    properties:
      interactions:
        items:
          $ref: '#/definitions/store.AgentInteractionRead'
        type: array
    type: object
...
  dev.resolveAgentInteractionBody:
    properties:
      answers:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      approved:
        type: boolean
      cancelled:
        type: boolean
      note:
        type: string
    type: object
  dev.resolveAgentInteractionResponse:
    properties:
      already_resolved:
        type: boolean
      applied:
        type: boolean
      interaction:
        $ref: '#/definitions/store.AgentInteractionRead'
    type: object
...
  store.AgentInteractionRead:
    properties:
      agent_name:
        type: string
      agent_run_id:
        type: string
      conversation_id:
        type: string
      conversation_title:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      kind:
        type: string
      request:
        additionalProperties: {}
        type: object
      request_id:
        type: string
      resolution_source:
        type: string
      resolved_actor:
        type: string
      resolved_at:
        type: string
      resolved_by:
        type: string
      response:
        additionalProperties: {}
        type: object
      status:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Parsar harness API. Optional Feishu SSO and event delivery
    are enabled only when the corresponding PARSAR_FEISHU_* values are configured.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Parsar API
  version: 0.1.0
paths:
...
  /api/v1/workspaces/{workspaceID}/interactions:
    get:
      description: Returns durable permission and AskUserQuestion requests. status
        accepts pending, decided, or expired.
      operationId: listAgentInteractions
      parameters:
      - description: Workspace UUID
        in: path
        name: workspaceID
        required: true
        type: string
      - description: pending, decided, or expired
        in: query
        name: status
        type: string
      - description: Maximum rows (1-200)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dev.listAgentInteractionsResponse'
        "400":
          description: Invalid status or limit
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Caller is not a workspace member
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Interaction store unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List workspace approval and user-question requests
      tags:
      - interactions
  /api/v1/workspaces/{workspaceID}/interactions/{interactionID}/resolve:
    post:
      consumes:
      - application/json
      description: Permission requests require approved. AskUserQuestion requests
        require an answers map keyed by stable question ID, or cancelled=true.
      operationId: resolveAgentInteraction
      parameters:
      - description: Workspace UUID
        in: path
        name: workspaceID
        required: true
        type: string
      - description: Interaction UUID
        in: path
        name: interactionID
        required: true
        type: string
      - description: Human decision
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dev.resolveAgentInteractionBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dev.resolveAgentInteractionResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Caller is not a writable workspace member
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Interaction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Interaction is currently being resolved
          schema:
            additionalProperties:
              type: string
            type: object
        "410":
          description: Interaction expired or runtime request ended
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Runtime temporarily unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Resolve a pending approval or user question
      tags:
      - interactions
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    in: cookie
    name: parsar_session
    type: apiKey
swagger: "2.0"